src/backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.5", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "destiny-backend"
version = "0.1.0"
description = "Backend for DEStiny Simulation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-ninja>=1.0.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.2.4",
    "gunicorn>=23.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pre-commit>=4.2.0",
    "django-cors-headers>=4.7.0",
    "whitenoise>=6.0.0",
    "destiny-sim",
]

[tool.uv.sources]
destiny-sim = { path = "../engine", editable = true }

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest-django>=4.11.1",
    "ruff>=0.11.13",
    "uvicorn>=0.34.3",
]

[tool.setuptools]
packages = ["simulation", "utils", "settings"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings.test_settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short --strict-markers"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = []
